requirements.txt
PyQt5==5.15.9
pyinstaller==6.5.0
Pillow==10.2.0 